requirements.txt
# pip requirements for running in a virtualenv

Django==1.3.1
Jinja2==2.6
M2Crypto==0.21.1
PIL==1.1.7
Paste==1.7.5.1
PyAMF==0.6.1
Pygments==1.4
South==0.7.3
Sphinx==1.1.2
amqplib==1.0.2
anyjson==0.3.1
boto==2.1.1
carrot==0.10.7
celery==2.4.5
decorator==3.3.2
django-celery==2.4.2
django-mailer==0.1.0
django-picklefield==0.1.9
django-storages==1.1.3
django-tagging==0.3.1
django-nose==0.1.3
docutils==0.8.1
-e git://github.com/TireSwingSoftware/fastenc.git@61cd5228b46b02b17fbf1e08207902974759f3cb#egg=fastenc-dev
# importlib only needed on python2.6 and earlier
importlib==1.0.2
kombu==1.5.1
mock==0.8.0
ordereddict==1.1
psycopg2==2.4.1
pycurl==7.19.0
pyparsing==1.5.6
python-dateutil==2.0
python-memcached==1.48
recaptcha-client==1.0.6
simplejson==2.3.0
wsgiref==0.1.2
